+++ packages/node/src/helpers/vk.ts
@@ -3,12 +3,12 @@ import { BaseHelper } from "./base";
 export default class VKHelper extends BaseHelper {
   // eslint-disable-next-line @typescript-eslint/require-await
   async getVideoId(url: URL) {
-    const pathID = /^\/(video|clip)-?\d{8,9}_\d{9}$/.exec(url.pathname);
+    const pathID = /^\/(video|clip)-?\d+_\d+$/.exec(url.pathname);
     if (pathID) {
       return pathID[0].slice(1);
     }
 
-    const idInsidePlaylist = /\/playlist\/[^/]+\/(video-?\d{8,9}_\d{9})/.exec(
+    const idInsidePlaylist = /\/playlist\/[^/]+\/(video-?\d+_\d+)/.exec(
       url.pathname,
     );
     if (idInsidePlaylist) {